requirements.txt
Flask
flask-cors
pymongo
python-dotenv
openai
python-docx
SpeechRecognition
pydub
werkzeug
